+++ packages/react/src/components/Card/CardToolbar.jsx
@@ -5,6 +5,7 @@ import { Close16, Popup16, Settings16 } from '@carbon/icons-react';
 import { OverflowMenu, OverflowMenuItem } from 'carbon-components-react';
 import classnames from 'classnames';
 import keyBy from 'lodash/keyBy';
+import { useLangDirection } from 'use-lang-direction';
 
 import { settings } from '../../constants/Settings';
 import { TimeRangeOptionsPropTypes } from '../../constants/CardPropTypes';
@@ -73,6 +74,7 @@ const propTypes = {
     closeLabel: PropTypes.string,
     expandLabel: PropTypes.string,
     settingsLabel: PropTypes.string,
+    extraActionLabel: PropTypes.string,
   }),
   testId: PropTypes.string,
   locale: PropTypes.string,
@@ -103,6 +105,7 @@ const defaultProps = {
     closeLabel: 'Close',
     expandLabel: 'Expand',
     settingsLabel: 'Settings',
+    extraActionLabel: 'Action Label',
   },
   testId: 'card-toolbar',
   dateTimeMask: 'YYYY-MM-DD HH:mm',
@@ -122,8 +125,11 @@ const CardToolbar = ({
   testId,
   locale,
   dateTimeMask,
+  extraActions,
 }) => {
   const mergedI18n = { ...defaultProps.i18n, ...i18n };
+  const langDir = useLangDirection();
+  const overflowMenuPosition = React.useMemo(() => langDir === 'ltr', [langDir]);
   // maps the timebox internal label to a translated string
   // Need the default here in case that the CardToolbar is used by multiple different components
   // Also needs to reassign itself if i18n changes
@@ -165,11 +171,43 @@ const CardToolbar = ({
     [onCardAction]
   );
 
+  const extraActionSlot =
+    extraActions && Object.keys(extraActions).length !== 0 ? (
+      extraActions.children && extraActions.children.length ? (
+        <OverflowMenu
+          flipped={overflowMenuPosition}
+          title={mergedI18n.overflowMenuDescription}
+          iconDescription={mergedI18n.overflowMenuDescription}
+        >
+          {extraActions.children.map((child, i) =>
+            !child.hidden ? (
+              <OverflowMenuItem
+                data-testid={`${testId}-extra-overflow-menu-item-${i}`}
+                key={`${child.id}-${i}`}
+                itemText={child.itemText}
+                disabled={child.disabled}
+                onClick={() => (child.callback ? child.callback(child) : null)}
+              />
+            ) : null
+          )}
+        </OverflowMenu>
+      ) : extraActions.icon ? (
+        <ToolbarSVGWrapper
+          title={mergedI18n.extraActionLabel}
+          onClick={() => (extraActions.callback ? extraActions.callback(extraActions) : null)}
+          iconDescription={mergedI18n.extraActionLabel}
+          renderIcon={extraActions.icon}
+          testId={`${testId}-extra-single-action`}
+          disabled={extraActions.disabled}
+        />
+      ) : null
+    ) : null;
+
   return isEditable ? (
     <div data-testid={testId} className={classnames(className, `${iotPrefix}--card--toolbar`)}>
       {(availableActions.clone || availableActions.delete) && (
         <OverflowMenu
-          flipped
+          flipped={overflowMenuPosition}
           title={mergedI18n.overflowMenuDescription}
           iconDescription={mergedI18n.overflowMenuDescription}
         >
@@ -248,13 +286,16 @@ const CardToolbar = ({
       {availableActions.expand ? (
         <>
           {isExpanded ? (
-            <ToolbarSVGWrapper
-              title={mergedI18n.closeLabel}
-              onClick={() => onCardAction(CARD_ACTIONS.CLOSE_EXPANDED_CARD)}
-              iconDescription={mergedI18n.closeLabel}
-              renderIcon={Close16}
-              testId={`${testId}-close-button`}
-            />
+            <>
+              {availableActions.extra ? extraActionSlot : null}
+              <ToolbarSVGWrapper
+                title={mergedI18n.closeLabel}
+                onClick={() => onCardAction(CARD_ACTIONS.CLOSE_EXPANDED_CARD)}
+                iconDescription={mergedI18n.closeLabel}
+                renderIcon={Close16}
+                testId={`${testId}-close-button`}
+              />
+            </>
           ) : (
             <ToolbarSVGWrapper
               title={mergedI18n.expandLabel}
@@ -268,6 +309,7 @@ const CardToolbar = ({
           )}
         </>
       ) : null}
+      {!isExpanded && availableActions.extra ? extraActionSlot : null}
     </div>
   );
 };