+++ packages/react/src/components/Table/StatefulTable.story.jsx
@@ -11,6 +11,7 @@ import StoryNotice from '../../internal/StoryNotice';
 import FlyoutMenu, { FlyoutMenuDirection } from '../FlyoutMenu/FlyoutMenu';
 import { csvDownloadHandler } from '../../utils/componentUtilityFunctions';
 import Button from '../Button/Button';
+import { DragAndDrop } from '../../utils/DragAndDropUtils';
 
 import StatefulTable from './StatefulTable';
 import {
@@ -118,7 +119,6 @@ export const StatefulTableWithNestedRowItems = (props) => {
           },
         }}
         actions={tableActions}
-        lightweight={boolean('Show an alternate header style (lightweight)', false)}
         {...props}
       />
     </div>
@@ -185,7 +185,6 @@ export const SimpleStatefulExample = () => {
       ])}
       style={{ maxWidth: select('table width', ['auto', '300px'], 'auto') }}
       useZebraStyles={boolean('Alternate colors in table rows (useZebraStyles)', false)}
-      lightweight={boolean('Show an alternate header style (lightweight)', false)}
       size={select(
         'Sets the height of the table rows (size)',
         ['xs', 'sm', 'md', 'lg', 'xl'],
@@ -387,7 +386,6 @@ export const StatefulExampleWithSingleNestedHierarchy = () => {
           ),
         }}
         actions={tableActions}
-        lightweight={boolean('Show an alternate header style (lightweight)', false)}
       />
     </div>
   );
@@ -459,7 +457,6 @@ export const SimpleStatefulExampleWithColumnOverflowMenu = () => {
           width: '150px',
         }))}
         actions={tableActions}
-        lightweight={boolean('Show an alternate header style (lightweight)', false)}
         size={select(
           'Sets the height of the table rows (size)',
           ['xs', 'sm', 'md', 'lg', 'xl'],
@@ -543,7 +540,6 @@ export const SimpleStatefulExampleWithAlignment = () => {
           'lg'
         )}
         actions={tableActions}
-        lightweight={boolean('Show an alternate header style (lightweight)', false)}
         options={{
           hasRowSelection: select(
             'Enable or Disable selecting single, multiple, or no rows (options.hasRowSelection)',
@@ -595,7 +591,6 @@ export const StatefulExampleWithEveryThirdRowUnselectable = () => {
         isSelectable: index % 3 !== 0,
       }))}
       actions={tableActions}
-      lightweight={boolean('Show an alternate header style (lightweight)', false)}
       size={select(
         'Sets the height of the table rows (size)',
         ['xs', 'sm', 'md', 'lg', 'xl'],
@@ -671,7 +666,6 @@ export const StatefulExampleWithExpansionMaxPagesAndColumnResize = () => {
             onDownloadCSV: (filteredData) => csvDownloadHandler(filteredData, 'my table data'),
           },
         }}
-        lightweight={boolean('Show an alternate header style (lightweight)', false)}
         size={select(
           'Sets the height of the table rows (size)',
           ['xs', 'sm', 'md', 'lg', 'xl'],
@@ -1155,7 +1149,6 @@ export const StatefulExampleWithCreateSaveViews = () => {
           ...tableActions,
           onUserViewModified,
         }}
-        lightweight={boolean('Show an alternate header style (lightweight)', false)}
         size={select(
           'Sets the height of the table rows (size)',
           ['xs', 'sm', 'md', 'lg', 'xl'],
@@ -1596,3 +1589,10 @@ export const WithMultiSorting = () => {
 };
 
 WithMultiSorting.storyName = 'with multi-sorting';
+WithMultiSorting.decorators = [
+  (Story) => (
+    <DragAndDrop>
+      <Story />
+    </DragAndDrop>
+  ),
+];