+++ packages/react/src/components/Table/StatefulTable.test.jsx
@@ -21,7 +21,7 @@ import {
 import { initialState, tableData } from './Table.story';
 import RowActionsCell from './TableBody/RowActionsCell/RowActionsCell';
 
-const { iotPrefix } = settings;
+const { prefix, iotPrefix } = settings;
 const mockActions = getMockActions(jest.fn);
 const selectData = getSelectData();
 const tableColumns = getTableColumns(selectData);
@@ -64,7 +64,7 @@ describe('stateful table with real reducer', () => {
         actions={mockActions}
       />
     );
-    statefulTable.find('button.bx--pagination__button--forward').simulate('click');
+    statefulTable.find(`button.${prefix}--pagination__button--forward`).simulate('click');
     expect(statefulTable.text()).toContain('100 of 100');
   });
   it('should show singleRowEditButtons when choosing to edit a row', async () => {
@@ -101,7 +101,7 @@ describe('stateful table with real reducer', () => {
       screen
         .getByText('whiteboard can eat 2')
         .closest('tr')
-        .querySelector('.bx--table-expand__button')
+        .querySelector(`.${prefix}--table-expand__button`)
     );
     expect(screen.getByText('whiteboard can eat 2A')).toBeTruthy();
     userEvent.click(screen.getByTestId(`${tableId}-row-2_A-row-actions-cell-overflow`));
@@ -115,19 +115,19 @@ describe('stateful table with real reducer', () => {
       screen
         .getByText('can pinocchio whiteboard 4')
         .closest('tr')
-        .querySelector('.bx--table-expand__button')
+        .querySelector(`.${prefix}--table-expand__button`)
     );
     userEvent.click(
       screen
         .getByText('can pinocchio whiteboard 4B')
         .closest('tr')
-        .querySelector('.bx--table-expand__button')
+        .querySelector(`.${prefix}--table-expand__button`)
     );
     userEvent.click(
       screen
         .getByText('can pinocchio whiteboard 4B-2')
         .closest('tr')
-        .querySelector('.bx--table-expand__button')
+        .querySelector(`.${prefix}--table-expand__button`)
     );
     expect(screen.getByText('can pinocchio whiteboard 4B-2-B')).toBeTruthy();
     userEvent.click(screen.getByTestId(`${tableId}-row-4_B-2-B-row-actions-cell-overflow`));