+++ packages/react/src/components/Table/Table.jsx
@@ -14,6 +14,7 @@ import { settings } from '../../constants/Settings';
 import FilterTags from '../FilterTags/FilterTags';
 import { RuleGroupPropType } from '../RuleBuilder/RuleBuilderPropTypes';
 import experimental from '../../internal/experimental';
+import deprecate from '../../internal/deprecate';
 
 import {
   TableColumnsPropTypes,
@@ -49,8 +50,11 @@ const propTypes = {
   tooltip: PropTypes.node,
   /** render zebra stripes or not */
   useZebraStyles: PropTypes.bool,
-  /**  lighter styling where regular table too visually heavy */
-  lightweight: PropTypes.bool,
+  /**  lighter styling where regular table too visually heavy. Deprecated. */
+  lightweight: deprecate(
+    PropTypes.bool,
+    `The 'lightweight' prop has been deprecated and will be removed in the next major version.`
+  ),
   /** Specify the properties of each column in the table */
   columns: TableColumnsPropTypes.isRequired,
   /** Specify the properties of each column group in the table. Defaults to empty column. */
@@ -234,6 +238,18 @@ const propTypes = {
       }),
       /* show the modal for selecting multi-sort columns */
       showMultiSortModal: PropTypes.bool,
+      multiSortModal: PropTypes.shape({
+        /**
+         * The anticipatedColumn is used to add the most recently click columnId to the UI of the
+         * MultiSort modal. This gives the user a better experience by pre-emptively adding the column
+         * they clicked multi-sort on to the multisort modal without changing state. They still have to
+         * click "Sort" to save it, or can click 'Cancel' or the 'X' to clear it.
+         */
+        anticipatedColumn: PropTypes.shape({
+          columnId: PropTypes.string,
+          direction: PropTypes.oneOf(['ASC', 'DESC']),
+        }),
+      }),
       /** Array with rowIds that are with loading active */
       loadingMoreIds: PropTypes.arrayOf(PropTypes.string),
     }),
@@ -319,7 +335,7 @@ export const defaultProps = (baseProps) => ({
   columnGroups: [],
   id: null,
   useZebraStyles: false,
-  lightweight: false,
+  lightweight: undefined,
   title: null,
   tooltip: null,
   secondaryTitle: null,
@@ -377,6 +393,8 @@ export const defaultProps = (baseProps) => ({
       },
       singleRowEditButtons: null,
       loadingMoreIds: [],
+      showMultiSortModal: false,
+      multiSortModal: undefined,
     },
   },
   actions: {
@@ -476,6 +494,7 @@ export const defaultProps = (baseProps) => ({
     multiSortAscending: 'Ascending',
     multiSortDescending: 'Descending',
     multiSortOverflowItem: 'Multi-sort',
+    multiSortDragHandle: 'Drag handle',
     // table error state
     tableErrorStateTitle: 'Unable to load the page',
     buttonLabelOnTableError: 'Refresh the page',
@@ -726,6 +745,20 @@ const Table = (props) => {
     );
   }
 
+  const multiSortColumns = useMemo(() => {
+    const arrayifiedSort = Array.isArray(view.table.sort)
+      ? view.table.sort
+      : view.table.sort !== undefined
+      ? [view.table.sort]
+      : [];
+
+    if (view.table.multiSortModal?.anticipatedColumn) {
+      return [...arrayifiedSort, view.table.multiSortModal.anticipatedColumn];
+    }
+
+    return arrayifiedSort;
+  }, [view.table.multiSortModal, view.table.sort]);
+
   return (
     <TableContainer
       style={style}
@@ -863,7 +896,16 @@ const Table = (props) => {
           </FilterTags>
         </section>
       ) : null}
-      <div className="addons-iot-table-container">
+      <div
+        className={classnames('addons-iot-table-container', {
+          // workaround hack to prevent double scrolling of the table and a filter dropdown
+          // because the Dropdown and Multiselect components don't support opening the menu
+          // items outside of the parent. This sets a minimum height for the table and applies
+          // a max-height to the dropdown list container based on that minimum height to prevent
+          // this issue.
+          [`${iotPrefix}-table-container--dropdown-height-fix`]: options.hasFilter,
+        })}
+      >
         <CarbonTable
           id={id}
           // TODO: remove id in v3
@@ -1087,7 +1129,7 @@ const Table = (props) => {
           testId={`${id}-multi-sort-modal`}
           columns={columns}
           ordering={view.table.ordering}
-          sort={Array.isArray(view.table.sort) ? view.table.sort : [view.table.sort]}
+          sort={multiSortColumns}
           actions={{
             ...pick(
               actions.table,